--- modulo5/aprofundamento-express/src/exe1.ts
@@ -1,6 +1,0 @@
-import { Request, Response } from "express";
-import { app } from "./index";
-
-app.get("/ping", (req:Request, res:Response) => {          
-    res.send("Pong! 🏓")
-})
