--- modulo5/aprofundamento-express/src/exe2.ts
@@ -1,6 +1,0 @@
-export type typeToDo ={
-    userId: number, 
-    id: number,
-    title: string,
-    completed: boolean
-}
